Empty return value added for rX()
When files call rX() from 'Communications_Format' it returns a value, and that value is stored in a local variable called 'rxResult'.  When the result has been seen and acted upon by the appropriate part of the code, it clears the 'rxResult' variable so it isn't acted upon twice.  Since the return value 'ACK' has the value '0x00', the absence of a return value, defined as 'EMPTY' has a value of '0x02' which is not used by any other return value.